langgraph-demo/pyproject.toml
[project]
name = "langgraph-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "langchain>=0.3.25",
    "langchain-core>=0.3.63",
    "langchain-mcp-adapters>=0.1.4",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "langgraph-supervisor>=0.0.27",
    "mcp>=1.9.2",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
]
